Don't export aggregate fields in Prometheus metrics
Exporting aggregate fields like "total" and "*" are an antipattern for
Prometheus metrics as it prevents the fields from being trivially
aggregatable. This leads to many mistakes when writing queries for
dashboards, etc.

Instead users should `sum without(field)(metric)` if the wish to get the
aggregate value for a particular metric.